modularise core styles
