+++ packages/core-ui/lib/core/Body/index.scss
@@ -1,0 +1,10 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.core-body {
+  position: relative;
+  display: flex;
+  flex-grow: 1;
+}
