+++ packages/core-ui/lib/core/ButtonRow/index.scss
@@ -1,0 +1,14 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+
+.core-button-row {
+  display: flex;
+  align-items: center;
+
+  &.y-margin {
+    margin-top: 1rem;
+  }
+}
