+++ packages/core-ui/lib/core/Main/index.scss
@@ -1,0 +1,12 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+
+.core-main {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+}
