+++ packages/core-ui/lib/core/Page/index.scss
@@ -1,0 +1,14 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles/variables" as v;
+@use "../styles";
+
+.core-page {
+  max-width: v.$page-max-width;
+  display: flex;
+  flex-direction: column;
+  padding-bottom: 4.5rem;
+  width: 100%;
+  margin: 0 auto;
+}
