+++ packages/core-ui/lib/core/Page/index.tsx
@@ -1,0 +1,19 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+import { ComponentBase } from "../../types";
+import "./index.scss";
+
+/**
+ * @name Page
+ * @summary
+ * A motion.div that wraps every page. Transitions can be applied to this wrapper that will affect
+ * the entire page.
+ */
+export const Page = ({ children, style }: ComponentBase) => {
+  return (
+    <div className="core-page" style={style}>
+      {children}
+    </div>
+  );
+};
