+++ packages/core-ui/lib/core/PageRow/index.scss
@@ -1,0 +1,18 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.core-page-row {
+  @include page-padding;
+
+  display: flex;
+  flex-shrink: 0;
+  flex-flow: row wrap;
+  width: 100%;
+  margin: 0;
+
+  &.y-margin {
+    margin: 1rem 0;
+  }
+}
