+++ packages/core-ui/lib/core/PageRow/index.tsx
@@ -1,0 +1,20 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+import { valEmpty } from "../../utils";
+import { RowProps } from "../types";
+import "./index.scss";
+
+/**
+ * @name PageRow
+ * @summary Used to separate page content based on rows. Commonly used with RowPrimary and
+ * RowSecondary.
+ */
+export const PageRow = ({ children, style, yMargin }: RowProps) => (
+  <div
+    className={`core-page-row${valEmpty(yMargin, "y-margin")}`}
+    style={style}
+  >
+    {children}
+  </div>
+);
