+++ packages/core-ui/lib/core/PageTitle/index.scss
@@ -1,0 +1,86 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles/variables" as v;
+@use "../styles";
+
+.core-page-title-hide-scrollable {
+  background: var(--background-default);
+  display: none;
+  position: fixed;
+  top: 0;
+  width: 100%;
+  height: 4rem;
+  z-index: 4;
+
+  @media (max-width: v.$page-width-medium-threshold) {
+    display: block;
+  }
+}
+
+.core-page-title {
+  @include page-padding;
+
+  background: var(--background-default);
+  display: flex;
+  flex-flow: column wrap;
+  justify-content: flex-end;
+  position: sticky;
+  margin-top: 5rem;
+  margin-bottom: 0.25rem;
+  padding-top: 0.5rem;
+  padding-bottom: 0;
+  top: 0;
+  transition: padding 0.3s ease-out;
+  width: 100%;
+  z-index: 5;
+
+  @media (max-width: v.$page-width-medium-threshold) {
+    top: 4rem;
+    padding-top: 0.75rem;
+    padding-bottom: 0.5rem;
+  }
+
+  .title {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    width: 100%;
+
+    > .right {
+      padding-left: 1rem;
+      flex-grow: 1;
+    }
+  }
+
+  h1 {
+    font-size: var(--title-font-size);
+    font-family: Unbounded, sans-serif;
+    font-weight: 700;
+    position: relative;
+    transform: scale(1);
+    left: 0;
+
+    @media (max-width: v.$page-width-medium-threshold) {
+      left: -1rem;
+      transform: scale(0.75);
+    }
+
+    transition: all var(--transition-duration);
+    margin: 0;
+  }
+
+  &.sticky {
+    padding-top: 1.5rem;
+    padding-bottom: 0.25rem;
+
+    .title {
+      margin-bottom: 0.75rem;
+
+      h1 {
+        transform: scale(0.75);
+        left: -1.25rem;
+      }
+    }
+  }
+}
