+++ packages/core-ui/lib/core/PageTitleTabs/index.scss
@@ -1,0 +1,32 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles/variables" as v;
+@use "../styles";
+
+.core-page-title-tabs {
+  max-width: v.$page-max-width;
+  border-bottom: 1px solid var(--border-primary-color);
+  overflow: hidden;
+  height: 3.6rem;
+  margin-top: 0.9rem;
+
+  @media (max-width: v.$page-width-medium-threshold) {
+    margin-top: 0.5rem;
+  }
+
+  > .scroll {
+    width: 100%;
+    height: 4.5rem;
+    overflow: auto hidden;
+  }
+
+  .inner {
+    display: flex;
+  }
+
+  &.sticky {
+    border-bottom: 0;
+    margin-top: 0.5rem;
+  }
+}
