+++ packages/core-ui/lib/core/Separator/index.scss
@@ -1,0 +1,10 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.core-separator {
+  border-bottom: 1px solid var(--border-primary-color);
+  width: 100%;
+  margin: 0.67rem 0;
+}
