+++ packages/core-ui/lib/core/Separator/index.tsx
@@ -1,0 +1,16 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+import { ComponentBase } from "../../types";
+import "./index.scss";
+
+/**
+ * @name Separator
+ * @summary A horizontal spacer with a bottom border. General spacer for separating content by
+ * row.
+ */
+export const Separator = ({ children, style }: ComponentBase) => (
+  <div className="core-separator" style={style}>
+    {children}
+  </div>
+);
