+++ packages/core-ui/lib/core/StatBoxRow/index.scss
@@ -1,0 +1,11 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.core-stat-box-row {
+  @include page-padding;
+
+  display: flex;
+  flex-direction: column;
+}
