+++ packages/core-ui/lib/core/Tx/index.scss
@@ -1,0 +1,84 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.core-tx {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+
+  &.margin {
+    margin-top: 1rem;
+  }
+
+  > .inner {
+    background: var(--background-modal-footer);
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    padding: 1rem;
+
+    > section {
+      width: 100%;
+      display: flex;
+      align-items: center;
+
+      p {
+        color: var(--text-color-secondary);
+        display: flex;
+        align-items: center;
+        font-size: 1rem;
+        margin: 0.1rem 0;
+        padding-left: 0.5rem;
+      }
+
+      > div {
+        display: flex;
+
+        &:first-child {
+          flex-direction: column;
+          justify-content: center;
+          flex-grow: 1;
+        }
+
+        &:last-child {
+          button {
+            margin-left: 0.75rem;
+          }
+        }
+      }
+    }
+  }
+
+  .sign {
+    display: flex;
+    align-items: center;
+    font-size: 0.9rem;
+    padding-bottom: 0.5rem;
+    margin: 0;
+
+    .badge {
+      border: 1px solid var(--border-secondary-color);
+      border-radius: 0.45rem;
+      padding: 0.2rem 0.5rem;
+      margin-right: 0.75rem;
+
+      > svg {
+        margin-right: 0.5rem;
+      }
+    }
+
+    .not-enough {
+      margin-left: 0.5rem;
+    }
+
+    .danger {
+      color: var(--status-danger-color);
+    }
+
+    > .icon {
+      margin-right: 0.3rem;
+    }
+  }
+}
