+++ packages/core-ui/lib/core/styles.scss
@@ -33,406 +33,3 @@ SPDX-License-Identifier: Apache-2.0 */
     }
   }
 }
-
-.core-entry {
-  background: var(--gradient-background);
-  width: 100%;
-  background-attachment: fixed;
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  flex-grow: 1;
-
-  h1,
-  h2,
-  h3,
-  input {
-    color: var(--text-color-primary);
-  }
-
-  h4,
-  h5,
-  p,
-  a {
-    color: var(--text-color-secondary);
-  }
-
-  path.primary,
-  ellipse.primary {
-    fill: var(--text-color-primary);
-  }
-
-  input {
-    padding: 0.7rem 0;
-    font-size: 1.1rem;
-    transition: all var(--transition-duration);
-
-    &::placeholder {
-      color: #aaa;
-    }
-  }
-
-  .textbox {
-    border-bottom: 1px solid #ddd;
-  }
-
-  .searchbox {
-    border: 1px solid #ddd;
-  }
-}
-
-.core-body {
-  position: relative;
-  display: flex;
-  flex-grow: 1;
-}
-
-.core-main {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-}
-
-.core-page {
-  max-width: v.$page-max-width;
-  display: flex;
-  flex-direction: column;
-  padding-bottom: 4.5rem;
-  width: 100%;
-  margin: 0 auto;
-}
-
-.core-side {
-  z-index: 7;
-  position: sticky;
-  top: 0;
-  height: 100vh;
-  flex: 0;
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-  transition: all 0.5s cubic-bezier(0.1, 1, 0.2, 1);
-
-  /* maximised by default, or minimised otherwise. */
-  min-width: v.$side-menu-maximised-width;
-  max-width: v.$side-menu-maximised-width;
-
-  &.minimised {
-    min-width: v.$side-menu-minimised-width;
-    max-width: v.$side-menu-minimised-width;
-  }
-
-  @media (max-width: v.$page-width-medium-threshold) {
-    position: fixed;
-    top: 0;
-    left: 0;
-
-    &.hidden {
-      left: (-(v.$side-menu-maximised-width));
-    }
-  }
-}
-
-.core-page-row {
-  @include page-padding;
-
-  display: flex;
-  flex-shrink: 0;
-  flex-flow: row wrap;
-  width: 100%;
-  margin: 0;
-
-  &.y-margin {
-    margin: 1rem 0;
-  }
-}
-
-.core-page-heading {
-  border-bottom: 1px solid var(--border-primary-color);
-  display: flex;
-  align-items: center;
-  flex-flow: row wrap;
-  padding-bottom: 0.75rem;
-  padding-top: 0.75rem;
-  margin-bottom: 0.25rem;
-  width: 100%;
-
-  > span {
-    margin-right: 1rem;
-  }
-
-  .right {
-    flex: 1 1 0%;
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: flex-end;
-
-    button {
-      margin: 0 0 0 1rem;
-    }
-  }
-}
-
-.core-page-title {
-  @include page-padding;
-
-  background: var(--background-default);
-  display: flex;
-  flex-flow: column wrap;
-  justify-content: flex-end;
-  position: sticky;
-  margin-top: 5rem;
-  margin-bottom: 0.25rem;
-  padding-top: 0.5rem;
-  padding-bottom: 0;
-  top: 0;
-  transition: padding 0.3s ease-out;
-  width: 100%;
-  z-index: 5;
-
-  @media (max-width: v.$page-width-medium-threshold) {
-    top: 4rem;
-    padding-top: 0.75rem;
-    padding-bottom: 0.5rem;
-  }
-
-  .title {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    width: 100%;
-
-    > .right {
-      padding-left: 1rem;
-      flex-grow: 1;
-    }
-  }
-
-  h1 {
-    font-size: var(--title-font-size);
-    font-family: Unbounded, sans-serif;
-    font-weight: 700;
-    position: relative;
-    transform: scale(1);
-    left: 0;
-
-    @media (max-width: v.$page-width-medium-threshold) {
-      left: -1rem;
-      transform: scale(0.75);
-    }
-
-    transition: all var(--transition-duration);
-    margin: 0;
-  }
-
-  &.sticky {
-    padding-top: 1.5rem;
-    padding-bottom: 0.25rem;
-
-    .title {
-      margin-bottom: 0.75rem;
-
-      h1 {
-        transform: scale(0.75);
-        left: -1.25rem;
-      }
-    }
-  }
-}
-
-.core-tx {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-
-  &.margin {
-    margin-top: 1rem;
-  }
-
-  > .inner {
-    background: var(--background-modal-footer);
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    padding: 1rem;
-
-    > section {
-      width: 100%;
-      display: flex;
-      align-items: center;
-
-      p {
-        color: var(--text-color-secondary);
-        display: flex;
-        align-items: center;
-        font-size: 1rem;
-        margin: 0.1rem 0;
-        padding-left: 0.5rem;
-      }
-
-      > div {
-        display: flex;
-
-        &:first-child {
-          flex-direction: column;
-          justify-content: center;
-          flex-grow: 1;
-        }
-
-        &:last-child {
-          button {
-            margin-left: 0.75rem;
-          }
-        }
-      }
-    }
-  }
-
-  .sign {
-    display: flex;
-    align-items: center;
-    font-size: 0.9rem;
-    padding-bottom: 0.5rem;
-    margin: 0;
-
-    .badge {
-      border: 1px solid var(--border-secondary-color);
-      border-radius: 0.45rem;
-      padding: 0.2rem 0.5rem;
-      margin-right: 0.75rem;
-
-      > svg {
-        margin-right: 0.5rem;
-      }
-    }
-
-    .not-enough {
-      margin-left: 0.5rem;
-    }
-
-    .danger {
-      color: var(--status-danger-color);
-    }
-
-    > .icon {
-      margin-right: 0.3rem;
-    }
-  }
-}
-
-.core-page-title-tabs {
-  max-width: v.$page-max-width;
-  border-bottom: 1px solid var(--border-primary-color);
-  overflow: hidden;
-  height: 3.6rem;
-  margin-top: 0.9rem;
-
-  @media (max-width: v.$page-width-medium-threshold) {
-    margin-top: 0.5rem;
-  }
-
-  > .scroll {
-    width: 100%;
-    height: 4.5rem;
-    overflow: auto hidden;
-  }
-
-  .inner {
-    display: flex;
-  }
-
-  &.sticky {
-    border-bottom: 0;
-    margin-top: 0.5rem;
-  }
-}
-
-.core-hide-scrollable {
-  background: var(--background-default);
-  display: none;
-  position: fixed;
-  top: 0;
-  width: 100%;
-  height: 4rem;
-  z-index: 4;
-
-  @media (max-width: v.$page-width-medium-threshold) {
-    display: block;
-  }
-}
-
-.core-separator {
-  border-bottom: 1px solid var(--border-primary-color);
-  width: 100%;
-  margin: 0.67rem 0;
-}
-
-.core-row-primary {
-  @include row-section-first;
-  @include row-section-last;
-
-  order: 0;
-  flex: 1;
-  flex-basis: 100%;
-  max-width: 100%;
-
-  @media (min-width: (v.$page-width-medium-threshold + 1px)) {
-    order: 0;
-    flex: 1;
-    flex-basis: 56%;
-    width: 56%;
-  }
-
-  @media (min-width: (v.$section-full-width-threshold + 400px)) {
-    flex-basis: 62%;
-    width: 62%;
-  }
-
-  &.v-last {
-    order: 1;
-  }
-}
-
-.core-row-secondary {
-  @include row-section-first;
-  @include row-section-last;
-
-  order: 0;
-  flex-basis: 100%;
-  width: 100%;
-  border-radius: 1rem;
-
-  &.v-last {
-    order: 1;
-  }
-
-  @media (min-width: (v.$page-width-medium-threshold + 1px)) {
-    order: 0;
-    flex: 1;
-    flex-basis: 44%;
-    width: 44%;
-    max-width: none;
-  }
-
-  @media (min-width: (v.$section-full-width-threshold + 400px)) {
-    flex-basis: 38%;
-    max-width: 38%;
-  }
-}
-
-.core-button-row {
-  display: flex;
-  align-items: center;
-
-  &.y-margin {
-    margin-top: 1rem;
-  }
-}
-
-.core-stat-box-row {
-  @include page-padding;
-
-  display: flex;
-  flex-direction: column;
-}
